+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   [`OptionWithValues.valueSplit`](https://ajalt.github.io/clikt/api/clikt/com.github.ajalt.clikt.parameters.options/-option-with-values/value-split.html).
 - Marking options as deprecated with [`option().deprecated()`](https://ajalt.github.io/clikt/api/clikt/com.github.ajalt.clikt.parameters.options/deprecated.html)
 - You can manually set the pattern to split envvars on by passing a pattern to the `envvarSplit` parameter of [`option()`](https://ajalt.github.io/clikt/api/clikt/com.github.ajalt.clikt.parameters.options/option.html)
+- [Option groups](https://ajalt.github.io/clikt/documenting/#grouping-options-in-help), [mutually exclusive groups](https://ajalt.github.io/clikt/options/#prompting-for-input), and [co-occurring groups](https://ajalt.github.io/clikt/options/#co-occurring-option-groups)
 
 ### Changed
 - If multiple `--` tokens are present on the command line, all subsequent occurrences after the first are now
@@ -19,6 +20,7 @@
   customizable. See [the docs](https://ajalt.github.io/clikt/documenting/) for more info, or the 
   [new sample](samples/ansicolors/README.md) for an example of customizing help output to use ANSI colors.
 - Some of the properties and constructor parameters for [`OptionWithValues`](https://ajalt.github.io/clikt/api/clikt/com.github.ajalt.clikt.parameters.options/-option-with-values/index.html) have changed.
+- The `OptionDelegate` interface has changed, and `GroupableOption` and `ParameterHolder` interfaces have been added to work with option groups.
 
 ## [1.7.0] - 2019-03-23
 ### Added